Move startUpload action to redux, remove the flux action file!
